Increase Git timeout in Jenkins jobs.
A full Git clone for some of our FreeBSD repositories is starting to
edge up against the default 10-minute timeout when run from Jenkins
(which runs at a lower priority than many other things). Bump this
timeout generously to 30 minutes (!) to ensure that we don't encounter
this problem again for a long time.